Finishing Up
I know this is some ugly spaghetti code, but I think it's good enough to
start testing with a live mic soon. I wouldn't really know where to
insert the right try/catch statements until I start having errors along
the way. So, for now, this is ready to go. Going to add some production
printing for monitoring purposes in the next commit.